LayoutTests/userscripts/user-script-video-document.html

<html>
<head>
<script>
if (window.layoutTestController) {
    window.layoutTestController.dumpAsText();
    window.layoutTestController.addUserScript("console.log('Start Script: ' + document.location.href)", true, true);
    window.layoutTestController.addUserScript("console.log('End Script: ' + document.location.href)", false, true);
}
</script>
</head>
<body>
<iframe src="resources/counting.mp4"></iframe>
</body>
</html>
