LayoutTests/userscripts/user-script-video-document-expected.txt

CONSOLE MESSAGE: line 1: End Script: user-script-video-document.html
CONSOLE MESSAGE: line 1: Start Script: counting.mp4
CONSOLE MESSAGE: line 1: End Script: counting.mp4
